feat(lobby): move the lobby creation form to a separate component
