+++ src/app/modules/lobby/components/all-lobbies/all-lobbies.component.html
@@ -5,6 +5,7 @@
 
 		<div class="tutorial-button-width" id="tutorial-create-new-multiplayer-lobby">
 			<button mat-raised-button [routerLink]="['create-lobby']" color="primary"><mat-icon>add</mat-icon> create new multiplayer lobby</button>
+			<button mat-raised-button [routerLink]="['join-lobby']" color="primary"><mat-icon>edit</mat-icon> join an existing multiplayer lobby</button>
 		</div>
 
 		<hr />
