+++ src/app/modules/lobby/lobby.module.ts
@@ -6,12 +6,14 @@ import { LobbyRoutingModule } from './lobby-routing.module';
 import { AllLobbiesComponent } from './components/all-lobbies/all-lobbies.component';
 import { CreateLobbyComponent } from './components/create-lobby/create-lobby.component';
 import { LobbyViewComponent } from './components/lobby-view/lobby-view.component';
+import { LobbyFormComponent } from './components/lobby-form/lobby-form.component';
 
 @NgModule({
 	declarations: [
 		AllLobbiesComponent,
 		CreateLobbyComponent,
-		LobbyViewComponent
+		LobbyViewComponent,
+		LobbyFormComponent
 	],
 	imports: [
 		CommonModule,
